+++ docs/.vitepress/theme/index.ts
@@ -5,6 +5,11 @@ import type { Theme } from 'vitepress';
 import { h } from 'vue';
 import DefaultTheme from 'vitepress/theme';
 import MiruSource from './components/MiruSource.vue';
+import MiruTable from './components/MiruTable.vue';
+
+import FilesizeDirective from '../../../src/directives/filesize';
+import OutsideDirective from '../../../src/directives/outside';
+import TooltipDirective from '../../../src/directives/tooltip';
 
 // https://vitepress.dev/guide/custom-theme
 export default {
@@ -15,6 +20,11 @@ export default {
         })
     },
     enhanceApp({ app, router, siteData }) {
+        app.directive('click-outside', OutsideDirective);
+        app.directive('filesize', FilesizeDirective);
+        app.directive('tooltip', TooltipDirective);
+
         app.component('MiruSource', MiruSource);
+        app.component('MiruTable', MiruTable);
     }
 } satisfies Theme