+++ histoire.setup.ts
@@ -4,6 +4,7 @@ import { defineSetupVue3 } from "@histoire/plugin-vue";
 
 import FilesizeDirective from './src/directives/filesize';
 import OutsideDirective from "./src/directives/outside";
+import TooltipDirective from './src/directives/tooltip';
 
 declare module "histoire" {
     interface CommonMeta {
@@ -12,6 +13,7 @@ declare module "histoire" {
 }
 
 export const setupVue3 = defineSetupVue3(({ app, addWrapper }) => {
-    app.directive('filesize', FilesizeDirective);
     app.directive('click-outside', OutsideDirective);
+    app.directive('filesize', FilesizeDirective);
+    app.directive('tooltip', TooltipDirective);
 });