+++ src/directives/outside.ts
@@ -1,11 +1,16 @@
 import { type Directive, type DirectiveBinding } from "vue";
 
+export interface CallbackItem {
+    cb: () => void;
+    once: boolean;
+}
+
 const listener = { 
     value: null,
     callbacks: new Map
 } as {
     value: null | ((event: Event) => void);
-    callbacks: Map<HTMLElement, () => void>;
+    callbacks: Map<HTMLElement, CallbackItem>;
 };
 
 /**
@@ -14,9 +19,12 @@ const listener = {
 function createDocumentListener() {
     listener.value = (event) => {
         let target = event.target as HTMLElement;
-        for (const [el, cb] of listener.callbacks.entries()) {
+        for (const [el, item] of listener.callbacks.entries()) {
             if (el && el.parentElement && el != target && !el.contains(target)) {
-                cb();
+                item.cb();
+            }
+            if (item.once) {
+                listener.callbacks.delete(el);
             }
         }
     };
@@ -37,9 +45,20 @@ const OutsideDirective: Directive = {
         if (listener.value == null) {
             createDocumentListener();
         }
+        const modifiers = Object.keys(bindings.modifiers);
+
+        // Parse Modifiers
+        let once = false;
+        for (const mod of modifiers) {
+            if (mod === 'once') {
+                once = true;
+            }
+        }
+
+        // Set Listener
         setTimeout(() => {
-            listener.callbacks.set(el, bindings.value);
-        }, 15);
+            listener.callbacks.set(el, { cb: bindings.value, once });
+        }, 10);
     },
 
     /**